PaymentServiceProvider: Move from ApiBootstrapResponse to ApiProvider
